+++ libs/designer/src/lib/ui/index.tsx
@@ -5,3 +5,4 @@ export * from './menuItems';
 export * from './templates';
 export * from './dialog/combineVariableDialog';
 export * from './dialog/triggerDescriptionDialog';
+export { getLogicAppsCategories } from './configuretemplate/resources';