+++ libs/logic-apps-shared/src/designer-client-services/lib/templateresource.ts
@@ -1,4 +1,4 @@
-import type { LogicAppsV2, Template } from '../../utils/src';
+import type { ArmResource, LogicAppsV2, Template } from '../../utils/src';
 import { AssertionErrorCode, AssertionException } from '../../utils/src';
 
 export interface WorkflowData {
@@ -7,6 +7,8 @@ export interface WorkflowData {
 }
 
 export interface ITemplateResourceService {
+  getTemplate: (id: string) => Promise<ArmResource<any>>;
+  getTemplateWorkflows: (id: string) => Promise<ArmResource<any>[]>;
   updateTemplate: (id: string, manifest: Template.TemplateManifest) => Promise<void>;
   updateWorkflow: (id: string, data: WorkflowData) => Promise<void>;
   createArtifact: (templateId: string, artifact: Template.Artifact) => Promise<void>;