+++ CHANGELOG.md
@@ -1,4 +1,31 @@
 # Logic Apps Designer
+## [5.68.0](https://github.com/Azure/LogicAppsUX/compare/v5.67.0...v5.68.0) (2025-03-30)
+
+
+### Features
+
+* **designer:** Add operation details in Input/Output channel ([#6909](https://github.com/Azure/LogicAppsUX/issues/6909)) ([4c15f26](https://github.com/Azure/LogicAppsUX/commit/4c15f263fd501404373d87bc68087a35f1d3fc38))
+
+## [5.67.0](https://github.com/Azure/LogicAppsUX/compare/v5.66.0...v5.67.0) (2025-03-30)
+
+
+### Features
+
+* **designer:** Entry point 3 for Agent Parameters ([#6905](https://github.com/Azure/LogicAppsUX/issues/6905)) ([c74f168](https://github.com/Azure/LogicAppsUX/commit/c74f168a61f969fcdebc73407a6d80410ea11917))
+* **designer:** Implement next link for agent actions repetitions ([#6901](https://github.com/Azure/LogicAppsUX/issues/6901)) ([66b9f4a](https://github.com/Azure/LogicAppsUX/commit/66b9f4aedc53a1d3e2d2d37cf2274a6603b3cdc4))
+* **designer:** Init channel tab for Agentic ([#6906](https://github.com/Azure/LogicAppsUX/issues/6906)) ([b44bb2c](https://github.com/Azure/LogicAppsUX/commit/b44bb2cc728b404063d8a843bc0a5bd5a324a065))
+* **Designer:** Making Agent Parameters Clickable ([#6894](https://github.com/Azure/LogicAppsUX/issues/6894)) ([7a18d94](https://github.com/Azure/LogicAppsUX/commit/7a18d949a333e69276e91af1710f837ce07cbf6a))
+* **designer:** Move chatbot in a standalone UI component ([#6902](https://github.com/Azure/LogicAppsUX/issues/6902)) ([a8c3fbe](https://github.com/Azure/LogicAppsUX/commit/a8c3fbede619abba8c133fc27d5310811622c768))
+* **template:** Adding connectors list in connections tab for custom templates ([#6903](https://github.com/Azure/LogicAppsUX/issues/6903)) ([5b110b8](https://github.com/Azure/LogicAppsUX/commit/5b110b8296fe24d0d3d4833f03c1fe099f31ac57))
+* **templates:** Adding template profile manifest form for custom template wizard ([#6907](https://github.com/Azure/LogicAppsUX/issues/6907)) ([f8d6d05](https://github.com/Azure/LogicAppsUX/commit/f8d6d051cd64f10b8c94189cd8b16cd913fd634e))
+* **Templates:** Converted `TemplateDisplay` description to markdown ([#6893](https://github.com/Azure/LogicAppsUX/issues/6893)) ([a028600](https://github.com/Azure/LogicAppsUX/commit/a0286003a7b0c02777a68832bad3a49402e947f7))
+
+
+### Bug Fixes
+
+* **templates:** Moving workflow metadata initialization in templatesdataprovider ([#6900](https://github.com/Azure/LogicAppsUX/issues/6900)) ([87324d2](https://github.com/Azure/LogicAppsUX/commit/87324d2d3c6e25f05fd13a685ab57eae0afbca49))
+* **vscode:** Update properties for agent actions response ([#6908](https://github.com/Azure/LogicAppsUX/issues/6908)) ([a385891](https://github.com/Azure/LogicAppsUX/commit/a3858916fdef83c212bb9eccb8a10216aac91937))
+
 ## [5.66.0](https://github.com/Azure/LogicAppsUX/compare/v5.65.0...v5.66.0) (2025-03-28)